Redesigning building layout (#98)

+++ server/src/services/property.service.ts
@@ -1,6 +1,7 @@
 import { Response } from 'express';
 import { PropertyRepository } from "../repositories";
 import { Property } from '../entity/Property';
+import { NumberOfAutoScalingGroups } from 'aws-sdk/clients/autoscaling';
 
 export class PropertyService {
 
@@ -11,7 +12,7 @@ export class PropertyService {
     }
 
 
-    public async createProperty(res: Response, tenantId: number, type: string, street: string, aptsuite: string, city: string, state: string, zipcode: string) {
+    public async createProperty(res: Response, tenantId: number, type: number, street: string, aptsuite: string, city: string, state: string, zipcode: string) {
         const propertyExists = await this.propertyRepository.findOne({TenantId: tenantId, Street: street, City: city, State: state});
         
         if (propertyExists) {
@@ -25,13 +26,14 @@ export class PropertyService {
 
 
     public async getProperties(tenantId: number) {
-        const properties = await this.propertyRepository.findAll();
+        console.log(this.propertyRepository)
+        const properties = await this.propertyRepository.find({ where: { TenantId: tenantId } });
         console.log(properties)
         return properties;
     }
 
 
-    public async updateProperty(res: Response, tenantId: number, id: number, type: string, street: string, aptsuite: string, city: string, state: string, zipcode: string) {
+    public async updateProperty(res: Response, tenantId: number, id: number, type: number, street: string, aptsuite: string, city: string, state: string, zipcode: string) {
         
         var property = await this.propertyRepository.findOne({Id: id});
 
@@ -52,20 +54,33 @@ export class PropertyService {
 
 
     public async getPropertiesByTenant(res: Response, tenantId: number) {
-        const properties = await this.propertyRepository.getPropertiesByTenant(tenantId);
+        const properties = await this.propertyRepository.find({ where: { TenantId: tenantId } });
         return res.status(200).json(properties);
     }
 
 
     public async getPropertyWithRelations(res: Response, tenantId: number, id: number) {
-        const property = await this.propertyRepository.getPropertyWithRelations(id);
+        const property = await this.propertyRepository.findOne({ where: {Id: id}, relations: ["Units"] });
         return res.status(200).json(property);
     }
 
 
     public async deleteProperty(res: Response, tenantId: number, id: number) {
-        const property = await this.propertyRepository.delete(id);
+        const propertyExists = await this.propertyRepository.find({TenantId: tenantId, Id: id});
+        
+        if (!propertyExists) {
+            return res.status(422).json({'errors': [{'msg': 'Property does not exist.'}]});
+        }
+
+        await this.propertyRepository.delete(id);
+
         return res.status(200).json({'msg': 'Removed property.'});
+    }
+
+    public async getPropertyTypes(res: Response, tenantId: number) {
+        const properties = await this.propertyRepository.getPropertyTypes(tenantId);
+        
+        return res.status(200).json(properties);
     }
 
 }
